Fix build on VS 16.9

+++ ME3ExplorerCore/UnrealScript/Compiling/BytecodeWriter.cs
@@ -20,7 +20,7 @@ namespace Unrealscript.Compiling
         {
             Pcc = pcc;
             Game = pcc.Game;
-            extNativeIndex = Game is MEGame.ME3 ? 0x70 : 0x60;
+            extNativeIndex = (byte) (Game is MEGame.ME3 ? 0x70 : 0x60);
         }
 
         public byte[] GetByteCode() => bytecode.ToArray();
